requirements.txt
Django==3.1
psycopg2-binary
Pillow
djangorestframework==3.11